Move verify_numericalized_example to torchtext_test_case
