+++ test/common/torchtext_test_case.py
@@ -65,3 +65,36 @@ class TorchtextTestCase(TestCase):
             test_numerical_features_dataset_file.write("0.2\t0\tteststring3\n")
             test_numerical_features_dataset_file.write("0.4\t12\tteststring4\n")
             test_numerical_features_dataset_file.write("0.9\t9\tteststring5\n")
+
+def verify_numericalized_example(field, test_example_data,
+                                 test_example_numericalized,
+                                 test_example_lengths=None,
+                                 batch_first=False, train=True):
+    """
+    Function to verify that numericalized example is correct
+    with respect to the Field's Vocab.
+    """
+    if isinstance(test_example_numericalized, tuple):
+        test_example_numericalized, lengths = test_example_numericalized
+        assert test_example_lengths == lengths.tolist()
+    if batch_first:
+        test_example_numericalized.data.t_()
+    # Transpose numericalized example so we can compare over batches
+    for example_idx, numericalized_single_example in enumerate(
+            test_example_numericalized.t()):
+        assert len(test_example_data[example_idx]) == len(numericalized_single_example)
+        assert numericalized_single_example.volatile is not train
+        for token_idx, numericalized_token in enumerate(
+                numericalized_single_example):
+            # Convert from Variable to int
+            numericalized_token = numericalized_token.data[0]
+            test_example_token = test_example_data[example_idx][token_idx]
+            # Check if the numericalized example is correct, taking into
+            # account unknown tokens.
+            if field.vocab.stoi[test_example_token] != 0:
+                # token is in-vocabulary
+                assert (field.vocab.itos[numericalized_token] ==
+                        test_example_token)
+            else:
+                # token is OOV and <unk> always has an index of 0
+                assert numericalized_token == 0
